moved create_project fixture down the line
